add some basic end to end functional tests
# Conflicts:
#	android-sdk/FlagshipSdk/src/main/java/com/flagship/sdk/core/models/FeatureFlagModels.kt

+++ android-sdk/FlagshipSdk/src/test/java/com/flagship/sdk/EdgeEvaluatorTest.kt
@@ -963,7 +963,7 @@ class EdgeEvaluatorTest {
             rolloutPercentage = rolloutPercentage,
             rules = rules,
             type = Type.Boolean,
-            updatedAt = System.currentTimeMillis().toDouble(),
+            updatedAt = System.currentTimeMillis(),
             variants = variants,
         )
 }
